[RFQ] Cancel, Accept, Quote WIP

+++ cluster-core/src/main/java/com/aeroncookbook/cluster/rfq/statemachine/states/RfqState.java
@@ -20,7 +20,7 @@ public interface RfqState
 {
     RfqStates getCurrentState();
 
-    int getCurrentStateId();
+    short getCurrentStateId();
 
     boolean canTransitionTo(RfqStates newState);
